// src/puzzle/skillTemplates.generated.ts
export const SKILL_PUZZLE_TEMPLATES: Record<string, RandomPuzzleTemplatePool> = {
  "cylinder-42": [
    {
      "arrows": [
        {
          "id": "arrow-1",
          "path": [
            [
              6,
              4,
              3
            ],
            [
              6,
              4,
              0
            ],
            [
              5,
              4,
              0
            ],
            [
              5,
              4,
              2
            ],
            [
              2,
              4,
              2
            ]
          ],
          "exitDirection": "-X",
          "color": 16039987,
          "lengthClass": "long"
        },
// ...
    },
// ...
};